test(preingestion-manager): move tests to from api-core crate
Refactor preingestion firmware and time-sync tests out of api-core
into preingestion-manager integration tests, sharing moved test
helpers and firmware test support.
