+++ crates/api-core/src/tests/host_bmc_firmware_test.rs
@@ -17,449 +17,36 @@
 
 use std::collections::HashMap;
 use std::fs;
-use std::net::{IpAddr, Ipv4Addr};
-use std::os::unix::fs::PermissionsExt;
-use std::str::FromStr;
 use std::time::Duration;
 
+use carbide_firmware::test_support::script_setup;
 use carbide_machine_controller::config::{FirmwareGlobal, TimePeriod};
 use carbide_machine_controller::handler::{
     MAX_FIRMWARE_UPGRADE_RETRIES, MAX_NEW_FIRMWARE_REPORTED_RESET_RETRIES,
 };
-use carbide_preingestion_manager::PreingestionManager;
-use carbide_redfish::libredfish::test_support::RedfishSimAction;
-use carbide_uuid::machine::MachineId;
 use common::api_fixtures::instance::TestInstance;
 use common::api_fixtures::{
     self, TestEnv, TestManagedHost, create_test_env_with_overrides, get_config,
 };
-use db::{self, DatabaseError};
 use model::firmware::{Firmware, FirmwareComponent, FirmwareComponentType, FirmwareEntry};
 use model::instance::status::tenant::TenantState;
 use model::machine::{HostReprovisionState, InstanceState, ManagedHostState};
 use model::machine_update_module::HOST_FW_UPDATE_HEALTH_REPORT_SOURCE;
-use model::site_explorer::{
-    Chassis, ComputerSystem, ComputerSystemAttributes, EndpointExplorationReport, EndpointType,
-    InitialBmcResetPhase, InitialResetPhase, Inventory, PowerDrainState, PowerState,
-    PreingestionState, Service, TimeSyncResetPhase,
-};
 use model::test_support::HardwareInfoTemplate;
 use regex::Regex;
 use rpc::forge::forge_server::Forge;
 use rpc::forge_agent_control_response::{Action, LegacyAction};
 use rpc::model::instance::snapshot::instance_snapshot_derive_status;
-use sqlx::PgConnection;
 use temp_dir::TempDir;
 use tokio::time::sleep;
 use tonic::Request;
 
 use crate::CarbideResult;
-use crate::cfg::file::CarbideConfig;
 use crate::machine_update_manager::MachineUpdateManager;
 use crate::tests::common;
 use crate::tests::common::api_fixtures::{
     TestEnvOverrides, create_managed_host_with_hardware_info_template, create_test_env,
 };
-use crate::tests::common::rpc_builder::DhcpDiscovery;
-
-#[crate::sqlx_test]
-async fn test_preingestion_bmc_upgrade(
-    pool: sqlx::PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let env = common::api_fixtures::create_test_env(pool.clone()).await;
-
-    let mgr = PreingestionManager::new(
-        pool.clone(),
-        env.config.preingestion_manager(),
-        env.redfish_sim.clone(),
-        env.test_meter.meter(),
-        None,
-        None,
-        None,
-        env.api.work_lock_manager_handle.clone(),
-        env.config.ntp_servers.clone(),
-    );
-
-    let mut txn = pool.begin().await.unwrap();
-
-    let response = env
-        .api
-        .discover_dhcp(
-            DhcpDiscovery::builder("b8:3f:d2:90:97:a6", "192.0.2.1")
-                .vendor_string("iDRac")
-                .tonic_request(),
-        )
-        .await?
-        .into_inner();
-
-    // First, a host where it's already up to date; it should go to complete
-    // after passing through the explicit NTP setup state.
-    let addr = response.address.as_str();
-    insert_endpoint_version(&mut txn, addr, "6.00.30.00", "1.13.2", false).await?;
-    txn.commit().await?;
-
-    mgr.run_single_iteration().await?;
-
-    mgr.run_single_iteration().await?;
-    let mut txn = pool.begin().await.unwrap();
-    assert!(
-        db::explored_endpoints::find_preingest_not_waiting_not_error(txn.as_mut())
-            .await?
-            .is_empty()
-    );
-    assert!(
-        db::explored_endpoints::find_all_preingestion_complete(&mut txn)
-            .await?
-            .len()
-            == 1
-    );
-
-    // Next, one that isn't up to date but it above preingestion limits.
-    db::explored_endpoints::delete(&mut txn, IpAddr::from_str(addr).unwrap()).await?;
-    insert_endpoint_version(&mut txn, addr, "5.1", "1.13.2", false).await?;
-    txn.commit().await?;
-    let mut txn = pool.begin().await.unwrap();
-
-    mgr.run_single_iteration().await?;
-    mgr.run_single_iteration().await?;
-
-    assert!(
-        db::explored_endpoints::find_preingest_not_waiting_not_error(txn.as_mut())
-            .await?
-            .is_empty()
-    );
-    assert!(
-        db::explored_endpoints::find_all_preingestion_complete(&mut txn)
-            .await?
-            .len()
-            == 1
-    );
-
-    // And now, one that's low enough to trigger preingestion upgrades.
-    db::explored_endpoints::delete(&mut txn, IpAddr::from_str(addr).unwrap()).await?;
-    insert_endpoint_version(&mut txn, addr, "4.9", "1.13.2", false).await?;
-    txn.commit().await?;
-
-    mgr.run_single_iteration().await?;
-    mgr.run_single_iteration().await?;
-    // The "upload" is synchronous now and will be complete at this point.
-
-    // At this point, we expect that it shows as having completed upload
-    let mut txn = pool.begin().await.unwrap();
-
-    let endpoints =
-        db::explored_endpoints::find_preingest_not_waiting_not_error(txn.as_mut()).await?;
-    assert!(endpoints.len() == 1);
-    let endpoint = endpoints.first().unwrap();
-    match &endpoint.preingestion_state {
-        // We expect it to be waiting for task completion
-        PreingestionState::UpgradeFirmwareWait {
-            task_id,
-            final_version,
-            upgrade_type,
-            ..
-        } => {
-            println!("Waiting on {task_id} {upgrade_type:?} {final_version}");
-        }
-        _ => {
-            panic!("Bad preingestion state: {endpoint:?}");
-        }
-    }
-    txn.commit().await?;
-
-    // Second firmware upload
-    mgr.run_single_iteration().await?;
-
-    let mut txn = pool.begin().await.unwrap();
-    let endpoints = db::explored_endpoints::find_all(txn.as_mut()).await?;
-    assert!(endpoints.len() == 1);
-    let endpoint = endpoints.first().unwrap();
-    if let PreingestionState::UpgradeFirmwareWait {
-        firmware_number, ..
-    } = endpoint.preingestion_state
-    {
-        assert_eq!(firmware_number, Some(1));
-    } else {
-        panic!("Bad preingestion state: {endpoint:?}");
-    };
-    txn.commit().await?;
-
-    // Let it go to NewFirmwareReportedWait
-    mgr.run_single_iteration().await?;
-
-    let mut txn = pool.begin().await.unwrap();
-    let endpoints = db::explored_endpoints::find_all(txn.as_mut()).await?;
-    assert!(endpoints.len() == 1);
-    let endpoint = endpoints.first().unwrap();
-    let PreingestionState::NewFirmwareReportedWait { .. } = endpoint.preingestion_state else {
-        panic!("Bad preingestion state: {endpoint:?}");
-    };
-    txn.commit().await?;
-
-    // One more, to make sure noething is weird with retrying resets
-    mgr.run_single_iteration().await?;
-
-    let mut txn = pool.begin().await.unwrap();
-    let endpoints = db::explored_endpoints::find_all(txn.as_mut()).await?;
-    assert!(endpoints.len() == 1);
-    let mut endpoint = endpoints.into_iter().next().unwrap();
-
-    // Now we simulate site explorer coming through and reading the new updated version
-    endpoint.report.service[0].inventories[0].version = Some("6.00.30.00".to_string());
-    assert!(
-        db::explored_endpoints::try_update(
-            endpoint.address,
-            endpoint.report_version,
-            &endpoint.report,
-            false,
-            &mut txn
-        )
-        .await?
-    );
-
-    txn.commit().await?;
-
-    // The next run of the state machine should see that the task shows as complete and move us back to checking again
-    mgr.run_single_iteration().await?;
-
-    let mut txn = pool.begin().await.unwrap();
-    let endpoints = db::explored_endpoints::find_all(txn.as_mut()).await?;
-    assert!(endpoints.len() == 1);
-    let endpoint = endpoints.first().unwrap();
-    match &endpoint.preingestion_state {
-        PreingestionState::RecheckVersions => {
-            println!("Rechecking versions");
-        }
-        _ => {
-            panic!("Bad preingestion state: {endpoint:?}");
-        }
-    }
-    txn.commit().await?;
-
-    // Now it should go to completion
-    mgr.run_single_iteration().await?;
-
-    let mut txn = pool.begin().await.unwrap();
-    assert!(
-        db::explored_endpoints::find_preingest_not_waiting_not_error(txn.as_mut())
-            .await?
-            .is_empty()
-    );
-    assert!(
-        db::explored_endpoints::find_all_preingestion_complete(&mut txn)
-            .await?
-            .len()
-            == 1
-    );
-    txn.commit().await?;
-
-    Ok(())
-}
-
-#[crate::sqlx_test]
-async fn test_preingestion_upgrade_script(
-    pool: sqlx::PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let (_tmpdir, config) = script_setup();
-    let env =
-        create_test_env_with_overrides(pool.clone(), TestEnvOverrides::with_config(config)).await;
-
-    let mgr = PreingestionManager::new(
-        pool.clone(),
-        env.config.preingestion_manager(),
-        env.redfish_sim.clone(),
-        env.test_meter.meter(),
-        None,
-        None,
-        None,
-        env.api.work_lock_manager_handle.clone(),
-        env.config.ntp_servers.clone(),
-    );
-
-    let response = env
-        .api
-        .discover_dhcp(
-            DhcpDiscovery::builder("b8:3f:d2:90:97:a6", "192.0.2.1")
-                .vendor_string("iDRac")
-                .tonic_request(),
-        )
-        .await?
-        .into_inner();
-
-    let addr = response.address.as_str();
-    let mut txn = pool.begin().await.unwrap();
-    db::explored_endpoints::delete(&mut txn, IpAddr::from_str(addr).unwrap()).await?;
-    insert_endpoint_version(&mut txn, addr, "0", "0", false).await?;
-    txn.commit().await?;
-
-    mgr.run_single_iteration().await?;
-    mgr.run_single_iteration().await?;
-
-    let mut txn = pool.begin().await.unwrap();
-    let endpoints =
-        db::explored_endpoints::find_preingest_not_waiting_not_error(txn.as_mut()).await?;
-    assert!(endpoints.len() == 1);
-    let endpoint = endpoints.first().unwrap();
-    match &endpoint.preingestion_state {
-        // We expect it to be waiting for task completion
-        PreingestionState::ScriptRunning => {}
-        _ => {
-            panic!("Bad preingestion state: {endpoint:?}");
-        }
-    }
-    txn.commit().await?;
-
-    tokio::time::sleep(std::time::Duration::from_secs(3)).await;
-
-    mgr.run_single_iteration().await?;
-
-    let mut txn = pool.begin().await.unwrap();
-    let endpoints =
-        db::explored_endpoints::find_preingest_not_waiting_not_error(txn.as_mut()).await?;
-    assert!(endpoints.len() == 1);
-    let endpoint = endpoints.first().unwrap();
-    match &endpoint.preingestion_state {
-        // We expect it to be have gone back to rechecking versions, we won't bother testing that here
-        PreingestionState::RecheckVersions => {}
-        _ => {
-            panic!("Bad preingestion state: {endpoint:?}");
-        }
-    }
-    txn.commit().await?;
-
-    Ok(())
-}
-
-async fn insert_endpoint_version(
-    txn: &mut PgConnection,
-    addr: &str,
-    bmc_version: &str,
-    uefi_version: &str,
-    powercycle_version: bool,
-) -> Result<(), DatabaseError> {
-    let model = if !powercycle_version {
-        "PowerEdge R750"
-    } else {
-        "Powercycle Test"
-    };
-    insert_endpoint(
-        txn,
-        addr,
-        "fm100hsag07peffp850l14kvmhrqjf9h6jslilfahaknhvb6sq786c0g3jg",
-        "Dell Inc.",
-        model,
-        bmc_version,
-        uefi_version,
-    )
-    .await
-}
-
-async fn insert_endpoint(
-    txn: &mut PgConnection,
-    addr: &str,
-    machine_id_str: &str,
-    vendor: &str,
-    model: &str,
-    bmc_version: &str,
-    uefi_version: &str,
-) -> Result<(), DatabaseError> {
-    let address = IpAddr::V4(Ipv4Addr::from_str(addr).unwrap());
-    db::explored_endpoints::insert(
-        address,
-        &build_exploration_report(vendor, model, bmc_version, uefi_version, machine_id_str),
-        false,
-        txn,
-    )
-    .await?;
-    db::explored_endpoints::set_preingestion_initial_bmc_reset(
-        address,
-        InitialBmcResetPhase::WaitForExplorerRefresh,
-        txn,
-    )
-    .await
-}
-
-fn build_exploration_report(
-    vendor: &str,
-    model: &str,
-    bmc_version: &str,
-    uefi_version: &str,
-    machine_id_str: &str,
-) -> EndpointExplorationReport {
-    let machine_id = if machine_id_str.is_empty() {
-        None
-    } else {
-        Some(MachineId::from_str(machine_id_str).unwrap())
-    };
-
-    let mut report = EndpointExplorationReport {
-        endpoint_type: EndpointType::Bmc,
-        vendor: Some(bmc_vendor::BMCVendor::Dell),
-        last_exploration_error: None,
-        last_exploration_latency: None,
-        managers: vec![],
-        systems: vec![ComputerSystem {
-            model: Some(model.to_string()),
-            ethernet_interfaces: vec![],
-            id: "".to_string(),
-            manufacturer: Some(vendor.to_string()),
-            serial_number: None,
-            attributes: ComputerSystemAttributes {
-                nic_mode: None,
-                is_infinite_boot_enabled: Some(true),
-            },
-            pcie_devices: vec![],
-            base_mac: None,
-            power_state: PowerState::On,
-            sku: None,
-            boot_order: None,
-        }],
-        chassis: vec![Chassis {
-            model: Some(model.to_string()),
-            id: "".to_string(),
-            manufacturer: Some(vendor.to_string()),
-            part_number: None,
-            serial_number: None,
-            network_adapters: vec![],
-            compute_tray_index: None,
-            physical_slot_number: None,
-            revision_id: None,
-            topology_id: None,
-        }],
-        service: vec![Service {
-            id: "".to_string(),
-            inventories: vec![
-                Inventory {
-                    id: "Installed-???__iDRAC.???".to_string(),
-                    description: None,
-                    version: Some(bmc_version.to_string()),
-                    release_date: None,
-                },
-                Inventory {
-                    id: "Current-159-1.13.2__BIOS.Setup.1-1".to_string(),
-                    description: None,
-                    version: Some(uefi_version.to_string()),
-                    release_date: None,
-                },
-            ],
-        }],
-        machine_id,
-        versions: HashMap::default(),
-        model: None,
-        machine_setup_status: None,
-        secure_boot_status: None,
-        lockdown_status: None,
-        power_shelf_id: None,
-        switch_id: None,
-        compute_tray_index: None,
-        physical_slot_number: None,
-        revision_id: None,
-        topology_id: None,
-        remediation_error: None,
-    };
-    report.model = report.model();
-    report
-}
 
 #[crate::sqlx_test]
 async fn test_postingestion_bmc_upgrade(pool: sqlx::PgPool) -> CarbideResult<()> {
@@ -966,207 +553,6 @@ fn test_merge_firmware_configs_write(
     let mut file = dir.clone();
     file.push("metadata.toml");
     fs::write(file, contents)?;
-
-    Ok(())
-}
-
-#[crate::sqlx_test]
-async fn test_preingestion_preupdate_powercycling(
-    pool: sqlx::PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let env = common::api_fixtures::create_test_env(pool.clone()).await;
-    tracing::debug!("{:?}", env.config.host_models);
-
-    let mgr = PreingestionManager::new(
-        pool.clone(),
-        env.config.preingestion_manager(),
-        env.redfish_sim.clone(),
-        env.test_meter.meter(),
-        None,
-        None,
-        None,
-        env.api.work_lock_manager_handle.clone(),
-        env.config.ntp_servers.clone(),
-    );
-
-    let mut txn = pool.begin().await.unwrap();
-
-    let response = env
-        .api
-        .discover_dhcp(
-            DhcpDiscovery::builder("b8:3f:d2:90:97:a6", "192.0.2.1")
-                .vendor_string("iDRac")
-                .tonic_request(),
-        )
-        .await?
-        .into_inner();
-
-    let addr = response.address.as_str();
-    insert_endpoint_version(&mut txn, addr, "4.9", "1.1", true).await?;
-    txn.commit().await?;
-
-    mgr.run_single_iteration().await?;
-    mgr.run_single_iteration().await?;
-    // The "upload" is synchronous now and will be complete at this point.
-
-    // Expect "reset" the BMC
-    let mut txn = pool.begin().await.unwrap();
-    let endpoints =
-        db::explored_endpoints::find_preingest_not_waiting_not_error(txn.as_mut()).await?;
-    let endpoint = endpoints.first().unwrap();
-    match &endpoint.preingestion_state {
-        PreingestionState::InitialReset { phase, .. } => {
-            assert_eq!(*phase, InitialResetPhase::BMCWasReset);
-        }
-        _ => {
-            panic!("Bad preingestion state: {:?}", endpoint.preingestion_state);
-        }
-    }
-    txn.commit().await?;
-    mgr.run_single_iteration().await?;
-
-    // Expect WaitHostBoot
-    let mut txn = pool.begin().await.unwrap();
-    let endpoints =
-        db::explored_endpoints::find_preingest_not_waiting_not_error(txn.as_mut()).await?;
-    let endpoint = endpoints.first().unwrap();
-    match &endpoint.preingestion_state {
-        PreingestionState::InitialReset { phase, .. } => {
-            assert_eq!(*phase, InitialResetPhase::WaitHostBoot);
-        }
-        _ => {
-            panic!("Bad preingestion state: {:?}", endpoint.preingestion_state);
-        }
-    }
-    // Pretend we waited
-    db::explored_endpoints::pregestion_hostboot_time_test(
-        IpAddr::V4(Ipv4Addr::from_str(addr).unwrap()),
-        &mut txn,
-    )
-    .await?;
-    txn.commit().await?;
-    mgr.run_single_iteration().await?;
-
-    // Recheck versions
-    let mut txn = pool.begin().await.unwrap();
-    let endpoints =
-        db::explored_endpoints::find_preingest_not_waiting_not_error(txn.as_mut()).await?;
-    let endpoint = endpoints.first().unwrap();
-    assert_eq!(
-        endpoint.preingestion_state,
-        PreingestionState::RecheckVersions
-    );
-    txn.commit().await?;
-    mgr.run_single_iteration().await?;
-
-    // At this point, we expect that it shows as having completed upload
-    let mut txn = pool.begin().await.unwrap();
-
-    let endpoints =
-        db::explored_endpoints::find_preingest_not_waiting_not_error(txn.as_mut()).await?;
-    assert!(endpoints.len() == 1);
-    let mut endpoint = endpoints.into_iter().next().unwrap();
-    match &endpoint.preingestion_state {
-        // We expect it to be waiting for task completion
-        PreingestionState::UpgradeFirmwareWait {
-            task_id,
-            final_version,
-            upgrade_type,
-            ..
-        } => {
-            println!("Waiting on {task_id} {upgrade_type:?} {final_version}");
-        }
-        _ => {
-            panic!("Bad preingestion state: {:?}", endpoint.preingestion_state);
-        }
-    }
-
-    // Now we simulate site explorer coming through and reading the new updated version
-    endpoint.report.service[0].inventories[0].version = Some("6.00.30.00".to_string());
-    assert!(
-        db::explored_endpoints::try_update(
-            endpoint.address,
-            endpoint.report_version,
-            &endpoint.report,
-            false,
-            &mut txn
-        )
-        .await?
-    );
-
-    txn.commit().await?;
-
-    for state in [
-        PowerDrainState::Off,
-        PowerDrainState::Powercycle,
-        PowerDrainState::On,
-        PowerDrainState::Off,
-        PowerDrainState::Powercycle,
-        PowerDrainState::On,
-    ] {
-        mgr.run_single_iteration().await?;
-
-        let mut txn = pool.begin().await.unwrap();
-        let endpoints = db::explored_endpoints::find_all(txn.as_mut()).await?;
-        assert!(endpoints.len() == 1);
-        let mut endpoint = endpoints.into_iter().next().unwrap();
-        tracing::debug!("State should be {state:?}");
-        match &endpoint.preingestion_state {
-            PreingestionState::ResetForNewFirmware {
-                delay_until,
-                last_power_drain_operation,
-                ..
-            } => {
-                assert!(delay_until.is_some());
-                assert_eq!(last_power_drain_operation.clone().unwrap(), state);
-                println!("Rechecking versions");
-            }
-            _ => {
-                panic!("Bad preingestion state: {:?}", endpoint.preingestion_state);
-            }
-        }
-
-        // At some point in here we would have picked up the new version
-        endpoint.report.service[0].inventories[1].version = Some("1.13.2".to_string());
-        assert!(
-            db::explored_endpoints::try_update(
-                endpoint.address,
-                endpoint.report_version,
-                &endpoint.report,
-                false,
-                &mut txn
-            )
-            .await?
-        );
-
-        txn.commit().await?;
-    }
-
-    mgr.run_single_iteration().await?;
-    let mut txn = pool.begin().await.unwrap();
-    let endpoints = db::explored_endpoints::find_all(txn.as_mut()).await?;
-    txn.commit().await?;
-    assert!(endpoints.len() == 1);
-    let endpoint = endpoints.first().unwrap();
-    let PreingestionState::RecheckVersions = endpoint.preingestion_state else {
-        panic!("Not in recheck versions: {:?}", endpoint.preingestion_state);
-    };
-
-    // Now it should go to completion
-    mgr.run_single_iteration().await?;
-    let mut txn = pool.begin().await.unwrap();
-    assert!(
-        db::explored_endpoints::find_preingest_not_waiting_not_error(txn.as_mut())
-            .await?
-            .is_empty()
-    );
-    assert!(
-        db::explored_endpoints::find_all_preingestion_complete(&mut txn)
-            .await?
-            .len()
-            == 1
-    );
-    txn.commit().await?;
 
     Ok(())
 }
@@ -1895,57 +1281,12 @@ async fn test_instance_upgrading_actual_part_2(
     Ok(())
 }
 
-fn script_setup() -> (TempDir, CarbideConfig) {
-    let tmpdir = TempDir::with_prefix("test_script_upgrade").unwrap();
-    let mut filename = tmpdir.path().to_path_buf();
-    filename.push("testscript_delete_me.sh");
-    fs::write(
-        &filename,
-        r#"#!/bin/bash
-
-echo BMC_IP $BMC_IP
-echo BMC_USERNAME $BMC_USERNAME
-echo BMC_PASSWORD $BMC_PASSWORD
-if ! echo $BMC_IP | grep -q ^192; then
-    echo "Wrong BMC IP"
-    exit 1
-fi
-sleep 2
-cat /proc/self/stat
-exit 0
-"#,
-    )
-    .unwrap();
-    fs::set_permissions(&filename, fs::Permissions::from_mode(0o755)).unwrap();
-
-    let mut config = get_config();
-    config.host_models = HashMap::from([(
-        "1".to_string(),
-        Firmware {
-            vendor: bmc_vendor::BMCVendor::Dell,
-            model: "PowerEdge R750".to_string(),
-            explicit_start_needed: false,
-            components: HashMap::from([(
-                FirmwareComponentType::Bmc,
-                FirmwareComponent {
-                    current_version_reported_as: Some(Regex::new("^Installed-.*__iDRAC.").unwrap()),
-                    preingest_upgrade_when_below: Some("1234".to_string()),
-                    known_firmware: vec![FirmwareEntry::standard_script(
-                        "1234",
-                        filename.to_str().unwrap(),
-                    )],
-                },
-            )]),
-            ordering: vec![FirmwareComponentType::Uefi, FirmwareComponentType::Bmc],
-        },
-    )]);
-
-    (tmpdir, config)
-}
-
 #[crate::sqlx_test]
 async fn test_script_upgrade(pool: sqlx::PgPool) -> CarbideResult<()> {
-    let (_tmpdir, config) = script_setup();
+    let (_tmpdir, host_models) = script_setup();
+    let mut config = get_config();
+    config.host_models = host_models;
+
     let env = create_test_env_with_overrides(pool, TestEnvOverrides::with_config(config)).await;
 
     let mh = common::api_fixtures::create_managed_host(&env).await;
@@ -2298,657 +1639,6 @@ async fn test_explicit_update(pool: sqlx::PgPool) -> CarbideResult<()> {
     };
 
     // That's sufficient to check the differences in this path
-    Ok(())
-}
-
-/// Test that when BMC time is in sync, preingestion proceeds normally with firmware checks
-#[crate::sqlx_test]
-async fn test_preingestion_time_sync_ok(
-    pool: sqlx::PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let mut config = get_config();
-    config.ntp_servers = vec![
-        "198.51.100.10".parse().unwrap(),
-        "198.51.100.11".parse().unwrap(),
-    ];
-    let env =
-        create_test_env_with_overrides(pool.clone(), TestEnvOverrides::with_config(config)).await;
-
-    let mgr = PreingestionManager::new(
-        pool.clone(),
-        env.config.preingestion_manager(),
-        env.redfish_sim.clone(),
-        env.test_meter.meter(),
-        None,
-        None,
-        None,
-        env.api.work_lock_manager_handle.clone(),
-        env.config.ntp_servers.clone(),
-    );
-
-    let mut txn = pool.begin().await.unwrap();
-
-    let response = env
-        .api
-        .discover_dhcp(
-            DhcpDiscovery::builder("b8:3f:d2:90:97:a6", "192.0.2.1")
-                .vendor_string("iDRac")
-                .tonic_request(),
-        )
-        .await?
-        .into_inner();
-
-    let addr = response.address.as_str();
-    let ip_addr = IpAddr::from_str(addr).unwrap();
-    // Insert endpoint with current versions that are up to date
-    insert_endpoint_version(&mut txn, addr, "6.00.30.00", "1.13.2", false).await?;
-    txn.commit().await?;
-
-    let timepoint = env.redfish_sim.timepoint();
-
-    // Run preingestion manager - should apply site NTP servers, check time sync,
-    // then check firmware and complete.
-    mgr.run_single_iteration().await?;
-
-    // Second iteration applies site NTP servers and records when that
-    // succeeded, but does not check BMC time until the convergence wait elapses.
-    mgr.run_single_iteration().await?;
-
-    let actions = env.redfish_sim.actions_since(&timepoint);
-    assert!(
-        actions
-            .all_hosts()
-            .iter()
-            .any(|a| matches!(a, RedfishSimAction::SetNtpServers(_))),
-        "Expected SetNtpServers when site NTP is configured"
-    );
-
-    let mut txn = pool.begin().await.unwrap();
-    let endpoints = db::explored_endpoints::find_all_by_ip(ip_addr, &mut txn).await?;
-    let endpoint = endpoints.first().expect("Endpoint should exist");
-    assert!(
-        matches!(
-            endpoint.preingestion_state,
-            PreingestionState::SetNtpServers {
-                set_at: Some(_),
-                attempts: 0
-            }
-        ),
-        "Expected SetNtpServers wait after applying NTP, got: {:?}",
-        endpoint.preingestion_state
-    );
-    txn.commit().await?;
-
-    // The next iteration should still wait for BMC NTP convergence.
-    mgr.run_single_iteration().await?;
-
-    let mut txn = pool.begin().await.unwrap();
-    db::explored_endpoints::set_preingestion_set_ntp_servers(
-        ip_addr,
-        Some(chrono::Utc::now() - chrono::TimeDelta::minutes(3)),
-        0,
-        &mut txn,
-    )
-    .await?;
-    txn.commit().await?;
-
-    // Once the convergence wait has elapsed, time sync and firmware checks complete.
-    mgr.run_single_iteration().await?;
-
-    let mut txn = pool.begin().await.unwrap();
-    // Should go directly to complete since time is in sync and firmware is up to date
-    assert!(
-        db::explored_endpoints::find_all_preingestion_complete(&mut txn)
-            .await?
-            .len()
-            == 1
-    );
-    txn.commit().await?;
-
-    Ok(())
-}
-
-/// Test that an empty NTP server config skips Redfish NTP setup and proceeds
-/// directly to initial checks from the SetNtpServers state.
-#[crate::sqlx_test]
-async fn test_preingestion_set_ntp_servers_empty(
-    pool: sqlx::PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let mut config = get_config();
-    config.ntp_servers.clear(); // Use empty NTP servers.
-
-    let env =
-        create_test_env_with_overrides(pool.clone(), TestEnvOverrides::with_config(config)).await;
-
-    let mgr = PreingestionManager::new(
-        pool.clone(),
-        env.config.preingestion_manager(),
-        env.redfish_sim.clone(),
-        env.test_meter.meter(),
-        None,
-        None,
-        None,
-        env.api.work_lock_manager_handle.clone(),
-        env.config.ntp_servers.clone(),
-    );
-
-    let response = env
-        .api
-        .discover_dhcp(
-            DhcpDiscovery::builder("b8:3f:d2:90:97:a6", "192.0.2.1")
-                .vendor_string("iDRac")
-                .tonic_request(),
-        )
-        .await?
-        .into_inner();
-
-    let addr = response.address.as_str();
-    let ip_addr = IpAddr::from_str(addr).unwrap();
-    let mut txn = pool.begin().await.unwrap();
-    insert_endpoint_version(&mut txn, addr, "6.00.30.00", "1.13.2", false).await?;
-
-    // Transition to start of SetNtpServers state with no attempts made so far.
-    db::explored_endpoints::set_preingestion_set_ntp_servers(ip_addr, None, 0, &mut txn).await?;
-    txn.commit().await?;
-
-    let timepoint = env.redfish_sim.timepoint();
-    mgr.run_single_iteration().await?;
-
-    // Expect no SetNtpServers actions since NTP server config is empty.
-    let actions = env.redfish_sim.actions_since(&timepoint);
-    assert!(
-        !actions
-            .all_hosts()
-            .iter()
-            .any(|a| matches!(a, RedfishSimAction::SetNtpServers(_))),
-        "Did not expect SetNtpServers when NTP server config is empty"
-    );
-
-    // Expect to go to complete since no need to set NTP server config.
-    let mut txn = pool.begin().await.unwrap();
-    assert_eq!(
-        db::explored_endpoints::find_all_preingestion_complete(&mut txn)
-            .await?
-            .len(),
-        1
-    );
-    txn.commit().await?;
-
-    Ok(())
-}
-
-/// Test that exhausting NTP setup attempts proceeds to initial checks without
-/// failing preingestion or trying to set NTP again.
-#[crate::sqlx_test]
-async fn test_preingestion_set_ntp_servers_max_attempts(
-    pool: sqlx::PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let mut config = get_config();
-    config.ntp_servers = vec!["198.51.100.10".parse().unwrap()];
-    let env =
-        create_test_env_with_overrides(pool.clone(), TestEnvOverrides::with_config(config)).await;
-
-    let mgr = PreingestionManager::new(
-        pool.clone(),
-        env.config.preingestion_manager(),
-        env.redfish_sim.clone(),
-        env.test_meter.meter(),
-        None,
-        None,
-        None,
-        env.api.work_lock_manager_handle.clone(),
-        env.config.ntp_servers.clone(),
-    );
-
-    let response = env
-        .api
-        .discover_dhcp(
-            DhcpDiscovery::builder("b8:3f:d2:90:97:a6", "192.0.2.1")
-                .vendor_string("iDRac")
-                .tonic_request(),
-        )
-        .await?
-        .into_inner();
-
-    let addr = response.address.as_str();
-    let ip_addr = IpAddr::from_str(addr).unwrap();
-    let mut txn = pool.begin().await.unwrap();
-    insert_endpoint_version(&mut txn, addr, "6.00.30.00", "1.13.2", false).await?;
-
-    db::explored_endpoints::set_preingestion_set_ntp_servers(ip_addr, None, 3, &mut txn).await?;
-    txn.commit().await?;
-
-    let timepoint = env.redfish_sim.timepoint();
-    mgr.run_single_iteration().await?;
-
-    let actions = env.redfish_sim.actions_since(&timepoint);
-    assert!(
-        !actions
-            .all_hosts()
-            .iter()
-            .any(|a| matches!(a, RedfishSimAction::SetNtpServers(_))),
-        "Did not expect SetNtpServers after max attempts are exhausted"
-    );
-
-    // The next iteration should go to complete since NTP setup is given up.
-    let mut txn = pool.begin().await.unwrap();
-    assert_eq!(
-        db::explored_endpoints::find_all_preingestion_complete(&mut txn)
-            .await?
-            .len(),
-        1
-    );
-    txn.commit().await?;
-
-    Ok(())
-}
-
-/// Test that preingestion handles the TimeSyncReset state machine correctly
-#[crate::sqlx_test]
-async fn test_preingestion_time_sync_reset_flow(
-    pool: sqlx::PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let env = common::api_fixtures::create_test_env(pool.clone()).await;
-
-    let mgr = PreingestionManager::new(
-        pool.clone(),
-        env.config.preingestion_manager(),
-        env.redfish_sim.clone(),
-        env.test_meter.meter(),
-        None,
-        None,
-        None,
-        env.api.work_lock_manager_handle.clone(),
-        env.config.ntp_servers.clone(),
-    );
-
-    let response = env
-        .api
-        .discover_dhcp(
-            DhcpDiscovery::builder("b8:3f:d2:90:97:a6", "192.0.2.1")
-                .vendor_string("iDRac")
-                .tonic_request(),
-        )
-        .await?
-        .into_inner();
-
-    let addr = response.address.as_str();
-    let ip_addr = IpAddr::from_str(addr).unwrap();
-
-    // Manually set up an endpoint in TimeSyncReset state to test the state machine
-    let mut txn = pool.begin().await.unwrap();
-    insert_endpoint_version(&mut txn, addr, "6.00.30.00", "1.13.2", false).await?;
-
-    // Set to TimeSyncReset Start phase
-    db::explored_endpoints::set_preingestion_time_sync_reset(
-        ip_addr,
-        TimeSyncResetPhase::Start,
-        0,
-        &mut txn,
-    )
-    .await?;
-    txn.commit().await?;
-
-    // Capture timepoint before running iteration
-    let timepoint = env.redfish_sim.timepoint();
-
-    // Run iteration - should initiate BMC reset and move to BMCWasReset
-    mgr.run_single_iteration().await?;
-
-    // Verify that SetUtcTimezone was called during the Start phase
-    let actions = env.redfish_sim.actions_since(&timepoint);
-    let all_actions = actions.all_hosts();
-    assert!(
-        all_actions.contains(&RedfishSimAction::SetUtcTimezone),
-        "Expected SetUtcTimezone action to be called during TimeSyncReset Start phase"
-    );
-
-    let mut txn = pool.begin().await.unwrap();
-    let endpoints =
-        db::explored_endpoints::find_preingest_not_waiting_not_error(txn.as_mut()).await?;
-    assert_eq!(endpoints.len(), 1);
-    let endpoint = endpoints.first().unwrap();
-    match &endpoint.preingestion_state {
-        PreingestionState::TimeSyncReset { phase, .. } => {
-            assert_eq!(*phase, TimeSyncResetPhase::BMCWasReset);
-        }
-        _ => {
-            panic!(
-                "Expected TimeSyncReset state, got: {:?}",
-                endpoint.preingestion_state
-            );
-        }
-    }
-    txn.commit().await?;
-
-    // Run iteration - should power on host and move to WaitHostBoot
-    mgr.run_single_iteration().await?;
-
-    let mut txn = pool.begin().await.unwrap();
-    let endpoints =
-        db::explored_endpoints::find_preingest_not_waiting_not_error(txn.as_mut()).await?;
-    let endpoint = endpoints.first().unwrap();
-    match &endpoint.preingestion_state {
-        PreingestionState::TimeSyncReset { phase, .. } => {
-            assert_eq!(*phase, TimeSyncResetPhase::WaitHostBoot);
-        }
-        _ => {
-            panic!(
-                "Expected TimeSyncReset WaitHostBoot, got: {:?}",
-                endpoint.preingestion_state
-            );
-        }
-    }
-
-    // Simulate time passage for host boot (pretend we waited 20 minutes)
-    db::explored_endpoints::pregestion_hostboot_time_test(ip_addr, &mut txn).await?;
-    txn.commit().await?;
-
-    // Run iteration - should check time sync again, and since mock BMC returns good time,
-    // proceed to check firmware versions which should complete since firmware is up-to-date
-    mgr.run_single_iteration().await?;
-
-    let mut txn = pool.begin().await.unwrap();
-    // After time sync reset completes and firmware check runs, endpoint should be in Complete state
-    // since the firmware versions are already up-to-date
-    let endpoints = db::explored_endpoints::find_all_by_ip(ip_addr, &mut txn).await?;
-    let endpoint = endpoints.first().expect("Endpoint should exist");
-    assert_eq!(
-        endpoint.preingestion_state,
-        PreingestionState::Complete,
-        "Expected Complete after successful time sync and firmware check, got: {:?}",
-        endpoint.preingestion_state
-    );
-    txn.commit().await?;
-
-    Ok(())
-}
-
-/// Test that when BMC time check returns an error, preingestion fails
-#[crate::sqlx_test]
-async fn test_preingestion_time_sync_check_error_fails(
-    pool: sqlx::PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    // Note: This test verifies the error handling path exists in the code.
-    // In practice, with a working mock BMC, this path might not be exercised.
-    // The actual behavior depends on whether the mock BMC's get_manager() method
-    // returns a valid DateTime or not.
-
-    let env = common::api_fixtures::create_test_env(pool.clone()).await;
-
-    let mgr = PreingestionManager::new(
-        pool.clone(),
-        env.config.preingestion_manager(),
-        env.redfish_sim.clone(),
-        env.test_meter.meter(),
-        None,
-        None,
-        None,
-        env.api.work_lock_manager_handle.clone(),
-        env.config.ntp_servers.clone(),
-    );
-
-    let response = env
-        .api
-        .discover_dhcp(
-            DhcpDiscovery::builder("b8:3f:d2:90:97:a6", "192.0.2.1")
-                .vendor_string("iDRac")
-                .tonic_request(),
-        )
-        .await?
-        .into_inner();
-
-    let addr = response.address.as_str();
-    let mut txn = pool.begin().await.unwrap();
-    insert_endpoint_version(&mut txn, addr, "6.00.30.00", "1.13.2", false).await?;
-    txn.commit().await?;
-
-    // Run preingestion - with mock BMC that has valid time, this should succeed
-    mgr.run_single_iteration().await?;
-
-    // The test passes if it doesn't panic - the mock BMC should return valid time
-    // and the endpoint should proceed to completion or firmware check
-    let mut txn = pool.begin().await.unwrap();
-    let endpoints = db::explored_endpoints::find_all(txn.as_mut()).await?;
-    assert_eq!(endpoints.len(), 1);
-    // Just verify we didn't fail - we should be in Complete or some valid state
-    let endpoint = &endpoints[0];
-    match &endpoint.preingestion_state {
-        PreingestionState::Failed { reason } => {
-            panic!("Unexpected failure: {}", reason);
-        }
-        _ => {
-            // Expected - time check passed or we're in a valid processing state
-        }
-    }
-    txn.commit().await?;
-
-    Ok(())
-}
-
-/// Test the retry logic when time sync fails after first reset attempt
-#[crate::sqlx_test]
-async fn test_preingestion_time_sync_retry_logic(
-    pool: sqlx::PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let env = common::api_fixtures::create_test_env(pool.clone()).await;
-
-    let mgr = PreingestionManager::new(
-        pool.clone(),
-        env.config.preingestion_manager(),
-        env.redfish_sim.clone(),
-        env.test_meter.meter(),
-        None,
-        None,
-        None,
-        env.api.work_lock_manager_handle.clone(),
-        env.config.ntp_servers.clone(),
-    );
-
-    let response = env
-        .api
-        .discover_dhcp(
-            DhcpDiscovery::builder("b8:3f:d2:90:97:a6", "192.0.2.1")
-                .vendor_string("iDRac")
-                .tonic_request(),
-        )
-        .await?
-        .into_inner();
-
-    let addr = response.address.as_str();
-    let ip_addr = IpAddr::from_str(addr).unwrap();
-
-    // Set up endpoint in TimeSyncReset WaitHostBoot phase
-    let mut txn = pool.begin().await.unwrap();
-    insert_endpoint_version(&mut txn, addr, "6.00.30.00", "1.13.2", false).await?;
-
-    // Manually set to WaitHostBoot phase as if we just finished a reset
-    db::explored_endpoints::set_preingestion_time_sync_reset(
-        ip_addr,
-        TimeSyncResetPhase::WaitHostBoot,
-        0,
-        &mut txn,
-    )
-    .await?;
-
-    // Simulate time has passed
-    db::explored_endpoints::pregestion_hostboot_time_test(ip_addr, &mut txn).await?;
-    txn.commit().await?;
-
-    // Run iteration - time check should pass (mock BMC returns valid time)
-    // and proceed to check firmware which should complete since firmware is up-to-date
-    mgr.run_single_iteration().await?;
-
-    let mut txn = pool.begin().await.unwrap();
-    // After time sync reset completes and firmware check runs, endpoint should be in Complete state
-    let endpoints = db::explored_endpoints::find_all_by_ip(ip_addr, &mut txn).await?;
-    let endpoint = endpoints.first().expect("Endpoint should exist");
-
-    // With a working mock BMC, time sync should succeed and firmware check should complete
-    match &endpoint.preingestion_state {
-        PreingestionState::Complete => {
-            // Expected - time sync passed and firmware is up-to-date
-        }
-        PreingestionState::RecheckVersions => {
-            // Could also be this if firmware check is still pending
-        }
-        PreingestionState::TimeSyncReset { phase, .. } => {
-            // If we're still in TimeSyncReset state, the reset is in progress
-            // But with mock BMC this shouldn't happen - we should have progressed
-            panic!(
-                "Unexpected: Still in TimeSyncReset state with phase {:?}",
-                phase
-            );
-        }
-        _ => {
-            // Could be other states if firmware upgrade is needed
-        }
-    }
-    txn.commit().await?;
-
-    Ok(())
-}
-
-/// When the BMC clock is still out of sync after a reset cycle but the retry
-/// budget is not yet exhausted, the endpoint should re-enter the reset cycle
-/// (TimeSyncReset Start) with an incremented attempt count rather than failing.
-#[crate::sqlx_test]
-async fn test_time_sync_retry_reenters_reset_before_failing(
-    pool: sqlx::PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let env = common::api_fixtures::create_test_env(pool.clone()).await;
-
-    // Simulate a BMC clock that is well past the 5 minute threshold.
-    env.redfish_sim.set_bmc_time_offset_seconds(600);
-
-    let mgr = PreingestionManager::new(
-        pool.clone(),
-        env.config.preingestion_manager(),
-        env.redfish_sim.clone(),
-        env.test_meter.meter(),
-        None,
-        None,
-        None,
-        env.api.work_lock_manager_handle.clone(),
-        env.config.ntp_servers.clone(),
-    );
-
-    let response = env
-        .api
-        .discover_dhcp(
-            DhcpDiscovery::builder("b8:3f:d2:90:97:a6", "192.0.2.1")
-                .vendor_string("iDRac")
-                .tonic_request(),
-        )
-        .await?
-        .into_inner();
-    let addr = response.address.as_str();
-    let ip_addr = IpAddr::from_str(addr).unwrap();
-
-    let mut txn = pool.begin().await.unwrap();
-    insert_endpoint_version(&mut txn, addr, "6.00.30.00", "1.13.2", false).await?;
-    // First reset cycle just finished (attempt 0), awaiting the boot-wait recheck.
-    db::explored_endpoints::set_preingestion_time_sync_reset(
-        ip_addr,
-        TimeSyncResetPhase::WaitHostBoot,
-        0,
-        &mut txn,
-    )
-    .await?;
-    // Backdate last_time so the boot wait is considered elapsed.
-    db::explored_endpoints::pregestion_hostboot_time_test(ip_addr, &mut txn).await?;
-    txn.commit().await?;
-
-    mgr.run_single_iteration().await?;
-
-    let mut txn = pool.begin().await.unwrap();
-    let endpoints = db::explored_endpoints::find_all_by_ip(ip_addr, &mut txn).await?;
-    match &endpoints
-        .first()
-        .expect("endpoint should exist")
-        .preingestion_state
-    {
-        PreingestionState::TimeSyncReset { phase, attempt, .. } => {
-            assert_eq!(
-                *phase,
-                TimeSyncResetPhase::Start,
-                "should retry reset cycle"
-            );
-            assert_eq!(*attempt, 1, "attempt counter should be incremented");
-        }
-        other => panic!("expected a retried TimeSyncReset, got: {other:?}"),
-    }
-    txn.commit().await?;
-
-    Ok(())
-}
-
-/// Once the reset retry budget is exhausted and the BMC clock is still out of
-/// sync, preingestion should fail terminally.
-#[crate::sqlx_test]
-async fn test_time_sync_fails_after_max_attempts(
-    pool: sqlx::PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let env = common::api_fixtures::create_test_env(pool.clone()).await;
-    env.redfish_sim.set_bmc_time_offset_seconds(600);
-
-    let mgr = PreingestionManager::new(
-        pool.clone(),
-        env.config.preingestion_manager(),
-        env.redfish_sim.clone(),
-        env.test_meter.meter(),
-        None,
-        None,
-        None,
-        env.api.work_lock_manager_handle.clone(),
-        env.config.ntp_servers.clone(),
-    );
-
-    let response = env
-        .api
-        .discover_dhcp(
-            DhcpDiscovery::builder("b8:3f:d2:90:97:a6", "192.0.2.1")
-                .vendor_string("iDRac")
-                .tonic_request(),
-        )
-        .await?
-        .into_inner();
-    let addr = response.address.as_str();
-    let ip_addr = IpAddr::from_str(addr).unwrap();
-
-    let mut txn = pool.begin().await.unwrap();
-    insert_endpoint_version(&mut txn, addr, "6.00.30.00", "1.13.2", false).await?;
-    // Final allowed reset cycle (attempt 2 == MAX_TIME_SYNC_RESET_ATTEMPTS - 1)
-    // just finished, awaiting the boot-wait recheck.
-    db::explored_endpoints::set_preingestion_time_sync_reset(
-        ip_addr,
-        TimeSyncResetPhase::WaitHostBoot,
-        2,
-        &mut txn,
-    )
-    .await?;
-    db::explored_endpoints::pregestion_hostboot_time_test(ip_addr, &mut txn).await?;
-    txn.commit().await?;
-
-    mgr.run_single_iteration().await?;
-
-    let mut txn = pool.begin().await.unwrap();
-    let endpoints = db::explored_endpoints::find_all_by_ip(ip_addr, &mut txn).await?;
-    match &endpoints
-        .first()
-        .expect("endpoint should exist")
-        .preingestion_state
-    {
-        PreingestionState::Failed { reason } => {
-            assert!(
-                reason.contains("time synchronization failed"),
-                "unexpected failure reason: {reason}"
-            );
-        }
-        other => panic!("expected Failed after exhausting retries, got: {other:?}"),
-    }
-    txn.commit().await?;
-
     Ok(())
 }
